refact: buiding CardListTitle component
