+++ src/components/molecules/index.ts
@@ -1,2 +1,3 @@
 export { VoteForm } from './VoteForm/VoteForm'
 export { Info } from './Info/Info'
+export { CardListTitle } from './CardListTitle/CardListTitle'
