+++ src/components/organisms/CardList/_cardList.sass
@@ -2,27 +2,6 @@
 
 .cardList
 
-  &__header
-    display: flex
-    justify-content: space-between
-    font-family: $ff-base
-
-    width: 95vw
-    margin: 0 1vw 0.7em 1.5vw
-
-    // &--list TODO: adjust margin-left
-    //   @media (min-width: $max-width-mobile and max-width: $max-width-tablet)
-
-    //   @media (min-width: $max-width-tablet)
-
-    &__title
-      font-size: $fz-h2-400-24
-      font-weight: 300
-      font-color: $clr-dark-00-2
-
-    &__selector
-      align-self: center
-
   &__list
     @include center-center
     padding: 0
